Resolved conflicts by:
- Accepting Docker/testcontainers configuration in test.yml
- Accepting DICOM implementation updates from main
- All 18 DICOM-related files updated with main branch versions

No FHIR-related conflicts - both branches made independent changes.

+++ dicom/pixel/lut.go
@@ -243,7 +243,7 @@ func ApplyModalityLUT(p *PixelData, slope, intercept float64) (*PixelData, error
 				if needsSigned {
 					data[i] = byte(int8(output))
 				} else {
-					data[i] = byte(uint8(output))
+					data[i] = byte(output)
 				}
 			} else {
 				// 16-bit output
